Add support to Google Text-to-Speech
- Reorganize TTS streamer class to support multiple TTS providers
  - Create an abstract class to act as an interface
  - Create a factory method with module loading support
- Implement the a streamer class for Google TTS using bi-directional streaming
