+++ README.md
@@ -19,4 +19,6 @@ Create a .env file with the following content:
 ```bash
 OPENAI_API_KEY = <openai_key>
 PORCUPINE_ACCESS_KEY = <porcupine_key>
+GOOGLE_PROJECT_ID = <google_project_id>
+GOOGLE_APPLICATION_CREDENTIALS = <google_credentials_path>
 ```
